+++ tests/local/test_ScotusParseEmailTest.py
@@ -1,6 +1,9 @@
-import os
+from pathlib import Path
 
-from juriscraper.scotus.scotus_email import SCOTUSEmail
+from juriscraper.scotus.scotus_email import (
+    SCOTUSEmail,
+    _SCOTUSConfirmationPageScraper,
+)
 from tests import TESTS_ROOT_EXAMPLES_SCOTUS
 from tests.local.PacerParseTestCase import PacerParseTestCase
 
@@ -10,9 +13,14 @@ class ScotusParseEmailTest(PacerParseTestCase):
 
     def setUp(self):
         self.maxDiff = 200000
+        self.path_root = Path(TESTS_ROOT_EXAMPLES_SCOTUS) / "email"
 
     def test_emails(self):
-        path_root = os.path.join(
-            TESTS_ROOT_EXAMPLES_SCOTUS, "dockets", "email"
+        self.parse_files(self.path_root, "*.eml", SCOTUSEmail)
+
+    def test_subscription(self):
+        self.parse_files(
+            self.path_root / "confirmation",
+            "*.html",
+            _SCOTUSConfirmationPageScraper,
         )
-        self.parse_files(path_root, "*.eml", SCOTUSEmail)